test/e2e/framework:move OpenWebSocketForURL to subpackage(new) websocket
