+++ test/e2e/common/BUILD
@@ -78,6 +78,7 @@ go_library(
         "//test/e2e/framework/rc:go_default_library",
         "//test/e2e/framework/skipper:go_default_library",
         "//test/e2e/framework/volume:go_default_library",
+        "//test/e2e/framework/websocket:go_default_library",
         "//test/utils:go_default_library",
         "//test/utils/image:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
