+++ test/e2e/framework/BUILD
@@ -78,7 +78,6 @@ go_library(
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/github.com/onsi/gomega/types:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
-        "//vendor/golang.org/x/net/websocket:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
         "//vendor/k8s.io/utils/exec:go_default_library",
     ],
@@ -143,6 +142,7 @@ filegroup(
         "//test/e2e/framework/testfiles:all-srcs",
         "//test/e2e/framework/timer:all-srcs",
         "//test/e2e/framework/volume:all-srcs",
+        "//test/e2e/framework/websocket:all-srcs",
     ],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
