+++ test/e2e/kubectl/BUILD
@@ -38,6 +38,7 @@ go_library(
         "//test/e2e/framework/pod:go_default_library",
         "//test/e2e/framework/service:go_default_library",
         "//test/e2e/framework/testfiles:go_default_library",
+        "//test/e2e/framework/websocket:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/utils:go_default_library",
         "//test/utils/crd:go_default_library",
